Add row creation func
Added row class with flex properties in order to apply it to the new
rows that are going to be created. Adjusted the sandbox flex properties
in order to make the grid centered. Added function for creating rows,
applying the row class to each in every iteration, calling the
createGridSquare function every iteration in a nested loop, and finally
appending the row to the sandbox every iteration