Add grid size change functionality
Added header and added new div to hold the button that will call the
function for creating a new grid. Adjusted styling of classes in order
to make everything on the page fit better and look nicer. Added function
that creates a button at the top of the page and calls other functions
in order to delete the current rows, get a number from the user, and
apply that number to the number of squares in the grid. Also created
the createNewRows and deleteRows helper functions so that the
createInputButton function would work as previously stated